% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{clean.shape}
\alias{clean.shape}
\title{clean.shape}
\usage{
clean.shape(x)
}
\arguments{
\item{x}{spatialPolygonsDataFrame}
}
\description{
Reprojects shape file to proj.env and turns factor data columns to strings.
}
\keyword{project,}
\keyword{spatialPolygonsDataFrame}
\keyword{transform,}
