% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{explode.shape.union}
\alias{explode.shape.union}
\title{explode.shape.union}
\usage{
explode.shape.union(b)
}
\arguments{
\item{b}{spatialPolygonsDataFrame}
}
\description{
When the spatialPolygon has only a length of 1, this will bring everything up so shapes are accessible.
}
\keyword{project,}
\keyword{spatialPolygonsDataFrame}
\keyword{transform,}
