% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{export.ogr}
\alias{export.ogr}
\title{export.ogr}
\usage{
export.ogr(shp, shp.name)
}
\arguments{
\item{shp}{spatialPolygonsDataFrame}

\item{shp.name}{output .shp file root name}
}
\description{
Simple function to export as an ogr file.
}
\keyword{export.ogr}
