% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{line.orthogonal}
\alias{line.orthogonal}
\title{line.orthogonal}
\usage{
line.orthogonal(midpoint, shp.line, proj.env.name = NULL)
}
\arguments{
\item{midpoint}{spatialPointsDataframe}

\item{shp.line}{spatialLinesDataframe}

\item{proj.env.name}{Projection envrionment name Example: denver}
}
\description{
Finds orthogonal line for the streetview direction aiming.
}
\keyword{line}
\keyword{orthogonal}
\keyword{streetview,}
