% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{projection.value}
\alias{projection.value}
\title{projection.value}
\usage{
projection.value(path.root = NULL, proj.name = NULL, list.only = FALSE)
}
\arguments{
\item{path.root}{root path Default: ~/Dropbox/pkg.data}

\item{proj.name}{projection name Example: 'wgs84'}

\item{list.only}{echos list of all possible projections}
}
\description{
Returns projection details from l.pkg
}
\keyword{name}
\keyword{projection}
