% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{shapes.check.identical}
\alias{shapes.check.identical}
\title{shapes.check.identical}
\usage{
shapes.check.identical(DT)
}
\arguments{
\item{DT}{SpatialPolygonsDataframe}
}
\description{
Check to see if there are any identical shapes and consolidate.
}
\keyword{consolidate}
\keyword{shapes}
