% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{shapes.extent}
\alias{shapes.extent}
\title{shapes.extent}
\usage{
shapes.extent(shp)
}
\arguments{
\item{shp}{spatialPolygonsDataFrame}
}
\description{
Returns the extent of any shapefile in WGS84 format.
}
\keyword{consolidate}
\keyword{shapes}
