% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{shapes.points.2.Buffer}
\alias{shapes.points.2.Buffer}
\title{shapes.points.2.buffer}
\usage{
shapes.points.2.Buffer(sp, shp)
}
\arguments{
\item{sp}{spatialPointsDataframe for lat/long}

\item{shp}{spatialPolygonsDataFrame the intersecting polygon}
}
\description{
Create a 1000 foot buffer around spatial points/parcel intersection this is currently only used for drug treatment locations.
}
\keyword{2}
\keyword{buffer}
\keyword{points}
