% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shapes.proj.env}
\alias{shapes.proj.env}
\title{shapes.proj.env}
\format{An object of class \code{"character"}.}
\source{
\href{'http://spatialreference.org/ref/epsg/2232/proj4js/'}{spatialreference.org proj4js}
}
\usage{
shapes.proj.env
}
\description{
The general projection environment for the Denver study area.
}
\examples{
cat(shapes.proj.env)
}
\references{
Project from spatial reference.org
(\href{'http://spatialreference.org/ref/epsg/2232/'}{spatailreference.org, epsg 2232})
}
\keyword{datasets}
