% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{shp.2.lines}
\alias{shp.2.lines}
\title{shp.2.lines}
\usage{
shp.2.lines(shp, proj.env.name = NULL)
}
\arguments{
\item{shp}{spatialPolygonsDataFrame}

\item{proj.env.name}{Projection envrionment name Example: denver}
}
\description{
Used by streetview to break apart a shapefile to its component outline.
}
\keyword{line}
\keyword{shape}
\keyword{shape.2.lines,}
\keyword{streetview,}
\keyword{to}
