% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{clean.shape}
\alias{clean.shape}
\title{clean.shape}
\usage{
clean.shape(x, proj.env.name = NULL)
}
\arguments{
\item{x}{spatialPolygonsDataFrame}

\item{proj.env.name}{projection environment Example: denver}
}
\description{
Reprojects shape file to proj.env and turns factor data columns to strings.
}
\keyword{project,}
\keyword{spatialPolygonsDataFrame}
\keyword{transform,}
