% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.shapes.R
\name{shapes.coords2points}
\alias{shapes.coords2points}
\title{shapes.coords2points}
\usage{
shapes.coords2points(DT, proj.env.name = NULL)
}
\arguments{
\item{DT}{data.table$long, data.table$lat}

\item{proj.env.name}{Projection envrionment name (Example: denver)}
}
\description{
Convert data table with lats and longs to spatial points.
}
\keyword{consolidate}
\keyword{shapes}
